Move qemu startup to its own script.
This causes us to reinstall and rebuild the image each time
we run qemu. This necessary now that we are loading files
from the disk because cmake makes it tricky (impossible?) to
have a command rely on a different target being "installed".

Because of this, updates to the teton executable were not being
picked up on the image.